training-pipeline/requirements.txt
azure-storage-blob
azure-identity
scikit-learn
scipy
matplotlib
pandas
azureml-core